+++ java-generator/it/src/it/escape-characters/src/test/java/io/fabric8/it/dummy/TestEscapeCharacters.java
@@ -19,11 +19,11 @@ import com.example.v1.Dummy;
 import com.example.v1.DummySpec;
 import com.example.v1.dummyspec.Package;
 import com.example.v1.dummyspec.package_.Foo;
-import com.fasterxml.jackson.databind.JsonNode;
 import io.fabric8.java.generator.testing.KubernetesResourceDiff;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import org.junit.jupiter.api.Test;
+import tools.jackson.databind.JsonNode;
 
 import java.nio.file.Files;
 import java.nio.file.Path;