+++ java-generator/it/src/it/ignore-unknown-toplevel-objects/src/test/java/io/fabric8/it/top/TestToplevelFields.java
@@ -15,7 +15,6 @@
  */
 package io.fabric8.it.top;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import io.argoproj.v1alpha1.Application;
 import io.argoproj.v1alpha1.ApplicationSpec;
 import io.argoproj.v1alpha1.application.Operation;
@@ -25,6 +24,7 @@ import io.fabric8.java.generator.testing.KubernetesResourceDiff;
 import io.fabric8.kubernetes.api.model.ObjectMeta;
 import io.fabric8.kubernetes.client.utils.Serialization;
 import org.junit.jupiter.api.Test;
+import tools.jackson.databind.JsonNode;
 
 import java.nio.file.Files;
 import java.nio.file.Path;